everything is calculating correctly and printing to log correctly. switching it over to lables.

+++ AOC2_Wk1/AOC2_Wk1/baseCustomCar.h
@@ -23,7 +23,7 @@ typedef enum {
 
 @property int costPerHour;
 @property NSArray *materials;
-@property int hoursPerJob;
+@property float hoursPerJob;
 @property NSString *jobDescription;
 @property int total;
